Refactor tally files
