+++ components/turf/pages/event/TallyEventRow.tsx
